Extract dialog into separate composable
